+++ examples/b.js
@@ -1,0 +1,4 @@
+function aaa( bbb ) {
+   }
+function ccc   (   ) {
+ddd();}