+++ examples/c.js
@@ -1,0 +1,7 @@
+function f()
+{
+  if (true)
+  {
+    g();
+  }
+}