+++ tests/lib/rules/multiline-block-statements.js
@@ -8,13 +8,11 @@ ruleTester.run('multiline-block-statements', rule, {
     'function f(){\na();\nb();\n}',
     'function f(){a();}',
   ],
-  invalid: [
-    {
-      code: 'function f(){a();b();}',
-      errors: [
-        { message: 'statement in a block to be on a new line.', type: 'BlockStatement' },
-        { message: 'statement in a block to be on a new line.', type: 'BlockStatement' },
-      ],
-    },
-  ],
+  invalid: [{
+    code: 'function f(){a();b();}',
+    errors: [
+      { message: 'Statement in a block must be on a new line.', type: 'BlockStatement' },
+      { message: 'Statement in a block must be on a new line.', type: 'BlockStatement' },
+    ],
+  }],
 });